[ML] Fix applying missing_bucket config to request payload.
