+++ x-pack/plugins/transform/public/app/common/index.ts
@@ -64,6 +64,7 @@ export {
   getPreviewTransformRequestBody,
   getCreateTransformRequestBody,
   getPivotQuery,
+  getRequestPayload,
   isDefaultQuery,
   isMatchAllQuery,
   isSimpleQuery,
