+++ x-pack/plugins/transform/public/app/common/request.ts
@@ -11,16 +11,29 @@ import { HttpFetchError } from '../../../../../../src/core/public';
 import type { IndexPattern } from '../../../../../../src/plugins/data/public';
 
 import type {
+  PivotTransformPreviewRequestSchema,
   PostTransformsPreviewRequestSchema,
   PutTransformsLatestRequestSchema,
   PutTransformsPivotRequestSchema,
   PutTransformsRequestSchema,
 } from '../../../common/api_schemas/transforms';
+import { DateHistogramAgg, HistogramAgg, TermsAgg } from '../../../common/types/pivot_group_by';
 
 import type { SavedSearchQuery } from '../hooks/use_search_items';
 import type { StepDefineExposedState } from '../sections/create_transform/components/step_define';
 import type { StepDetailsExposedState } from '../sections/create_transform/components/step_details';
 import { isPopulatedObject } from './utils/object_utils';
+
+import {
+  getEsAggFromAggConfig,
+  getEsAggFromGroupByConfig,
+  isGroupByDateHistogram,
+  isGroupByHistogram,
+  isGroupByTerms,
+  GroupByConfigWithUiSupport,
+  PivotAggsConfig,
+  PivotGroupByConfig,
+} from './';
 
 export interface SimpleQuery {
   query_string: {
@@ -80,6 +93,66 @@ export function getCombinedRuntimeMappings(
   }
   return undefined;
 }
+
+export const getMissingBucketConfig = (
+  g: GroupByConfigWithUiSupport
+): { missing_bucket?: boolean } => {
+  return g.missing_bucket !== undefined ? { missing_bucket: g.missing_bucket } : {};
+};
+
+export const getRequestPayload = (
+  pivotAggsArr: PivotAggsConfig[],
+  pivotGroupByArr: PivotGroupByConfig[]
+) => {
+  const request = {
+    pivot: {
+      group_by: {},
+      aggregations: {},
+    } as PivotTransformPreviewRequestSchema['pivot'],
+  };
+
+  pivotGroupByArr.forEach((g) => {
+    if (isGroupByTerms(g)) {
+      const termsAgg: TermsAgg = {
+        terms: {
+          field: g.field,
+          ...getMissingBucketConfig(g),
+        },
+      };
+      request.pivot.group_by[g.aggName] = termsAgg;
+    } else if (isGroupByHistogram(g)) {
+      const histogramAgg: HistogramAgg = {
+        histogram: {
+          field: g.field,
+          interval: g.interval,
+          ...getMissingBucketConfig(g),
+        },
+      };
+      request.pivot.group_by[g.aggName] = histogramAgg;
+    } else if (isGroupByDateHistogram(g)) {
+      const dateHistogramAgg: DateHistogramAgg = {
+        date_histogram: {
+          field: g.field,
+          calendar_interval: g.calendar_interval,
+          ...getMissingBucketConfig(g),
+        },
+      };
+      request.pivot.group_by[g.aggName] = dateHistogramAgg;
+    } else {
+      request.pivot.group_by[g.aggName] = getEsAggFromGroupByConfig(g);
+    }
+  });
+
+  pivotAggsArr.forEach((agg) => {
+    const result = getEsAggFromAggConfig(agg);
+    if (result === null) {
+      return;
+    }
+    request.pivot.aggregations[agg.aggName] = result;
+  });
+
+  return request;
+};
 
 export function getPreviewTransformRequestBody(
   indexPatternTitle: IndexPattern['title'],
